% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{areUniqueValues}
\alias{areUniqueValues}
\title{Are Unique Values}
\usage{
areUniqueValues(data)
}
\arguments{
\item{data:}{vector contains the values that we need to validate}
}
\value{
TRUE TRUE if the values are unique, and FALSE otherwise.
}
\description{
Check the uniquness of the values
}
\details{
To check the uniqueness of the values in a vector.
}

