% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{areValuesExisted}
\alias{areValuesExisted}
\title{Are the values of one vector contained in another vector}
\usage{
areValuesExisted(foriegnKeys, primaryKeys)
}
\arguments{
\item{foreignKeys:}{vector contains the values that we need to validation}

\item{primaryKeys:}{vector contains the values that we need to validation throw out}
}
\value{
TRUE if all the values of the first vector are included in the second vector, and FALSE in the otherwise case.
}
\description{
Use this function to validation that the values of one vector are already existed in another vector. If this function returns FALSE, you can use the function notExistedValues to see which values are existed in the first vector but not in the second vector
}
\details{
It is like foreign key/primary key validation in relational database.
}

