% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{correctMissingLeftZeroForTime}
\alias{correctMissingLeftZeroForTime}
\title{Correct the missing zero at the left of a time value}
\usage{
correctMissingLeftZeroForTime(time)
}
\arguments{
\item{time:}{a vector of integer/string values}
}
\value{
a vector of CHARACTORS contain the values with the correct format.
}
\description{
There are some time values have this format 921 which should be 0921 but because it is integer values, the zero on the left will not appear. This function is to handle these cases
}
\details{
Input is a vector of integer/string values
}

