\name{descopl-package}
\alias{descopl-package}
\alias{descopl}
\docType{package}
\title{
\packageTitle{descopl}
}
\description{
\packageDescription{descopl}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{descopl}
\packageIndices{descopl}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{descopl}

Maintainer: \packageMaintainer{descopl}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
