% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{doesItContainNullValues}
\alias{doesItContainNullValues}
\title{Does it contain null values}
\usage{
doesItContainNullValues(data)
}
\arguments{
\item{data:}{vector to be checked}
}
\description{
check if all the values in the vector are not null
}

