% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{isCorrectedDate}
\alias{isCorrectedDate}
\title{Is Correct Date}
\usage{
isCorrectedDate(date)
}
\arguments{
\item{date:}{a vector of integer/string values}
}
\value{
TRUE if all the values of the vector have the correct format, and otherwise it returns FALSE
}
\description{
Check if all the values of the input vectors have this format %Y%m%d
}
\details{
Input is a vector of integer/string values
}

