% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{isCorrectTime}
\alias{isCorrectTime}
\title{Is Correct Time}
\usage{
isCorrectTime(time)
}
\arguments{
\item{time:}{a vector of integer/string values}
}
\value{
TRUE if all the values of the vector have the correct format, and otherwise it returns FALSE
}
\description{
Check if all the values of the input vectors have this format %H%M
}
\details{
Input is a vector of integer/string values to be validated
}

