% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{isTheLengthOfAllTimeValuesThree}
\alias{isTheLengthOfAllTimeValuesThree}
\title{Is the length of all time values equal to three}
\usage{
isTheLengthOfAllTimeValuesThree(time)
}
\arguments{
\item{time:}{a vector of integer/string values}
}
\value{
TRUE when the lenght of all the values equal to three, and FALSE in the otherwise
}
\description{
Check if the length of all the values equal to three, we need that in order to know situations where the the leff zero is missing in the time value
}
\details{
Input is a vector of integer/string values
}

