% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{notExistedValues}
\alias{notExistedValues}
\title{Not existed values}
\usage{
notExistedValues(foreignKeys, primaryKeys)
}
\arguments{
\item{foreignKeys:}{vector contains the values that we need to validation}

\item{primaryKeys:}{vector contains the values that we need to validation throw out}
}
\description{
This function returns the values that exist in the first vector but don't exist in the second one
}

