% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{wrongDateValues}
\alias{wrongDateValues}
\title{Wrong Date Values}
\usage{
wrongDateValues(date)
}
\arguments{
\item{date:}{a vector of integer/string values}
}
\value{
a data frame contains two columns; WrongValues is a vector contains the wrong values, and WrongIndexes contains the indexes of the wrong values
}
\description{
Returns the wrong date values and their indexes.
}
\details{
Input is a vector of integer/string values
}

