% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{wrongTimeValues}
\alias{wrongTimeValues}
\title{Wrong Time Values}
\usage{
wrongTimeValues(time)
}
\arguments{
\item{time:}{a vector of integer/string values}
}
\value{
a data frame contains two columns; WrongValues is a vector contains the wrong values, and WrongIndexes contains the indexes of the wrong values
}
\description{
Returns the wrong time values and their indexes.
}
\details{
Input is a vector of integer/string values
}

