% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{notExistedValues}
\alias{notExistedValues}
\title{Not existed values}
\usage{
notExistedValues(foreignKeys, primaryKeys)
}
\arguments{
\item{foreignKeys:}{vector contains the values that we need to validation}

\item{primaryKeys:}{vector contains the values that we need to validation throw out}
}
\description{
This function returns a data frame contains the wrong values, whih are the values the exist in the first vector but do not exit in the second vector, and wrong indexes, which are the indexes of the wrong values
}

